requirement.txt
django
django-rest-framework
django-cors-headers
djangorestframework_simplejwt
factory_boy
